# Reduce memory allocations in interpolation