When using [venom](https://github.com/ovh/venom) I found the performance to be not as good as I hoped and used [pprof](https://go.dev/blog/pprof) to try to find some places where it can be improved. Venom seems to use `interpolate.Do` a lot, so I looked if there's anything to improve and found something that seems safe and doesn't make the code more complex:

Instead of allocating new maps, the existing ones can be reused. Deleting the keys while ranging over a map is safe in Go, see the section in [Effective Go](https://golang.org/doc/effective_go#for):

> ```go
> for key := range m {
>     if key.expired() {
>         delete(m, key)
>     }
> }
> ```

Previous benchmark:
```
$ go test -run=XXX -bench=. -benchmem
goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
BenchmarkDo-8          	    6561	    154820 ns/op	  115794 B/op	    1176 allocs/op
BenchmarkDoNothing-8   	   68073	     18401 ns/op	   10458 B/op	     119 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	2.824s
$ go test -run=XXX -bench=. -benchmem
goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
BenchmarkDo-8          	    7088	    149979 ns/op	  115689 B/op	    1176 allocs/op
BenchmarkDoNothing-8   	   66132	     18032 ns/op	   10456 B/op	     119 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	2.582s
```

After the change:
```
$ go test -run=XXX -bench=. -benchmem
goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
BenchmarkDo-8          	    7345	    141649 ns/op	   96200 B/op	    1161 allocs/op
BenchmarkDoNothing-8   	   64874	     18045 ns/op	   10456 B/op	     119 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	2.614s
$ go test -run=XXX -bench=. -benchmem
goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
BenchmarkDo-8          	    7208	    141367 ns/op	   96059 B/op	    1161 allocs/op
BenchmarkDoNothing-8   	   65907	     17964 ns/op	   10458 B/op	     119 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	2.527s
```

I haven't measured the difference when running a Venom scenario and it's probably really small, but I thought it's worth improving anyway.

@ovh/cds